tests/wpt/mozilla/tests/mozilla/document-characterSet-long.html

<!DOCTYPE html>
<meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
    assert_equals(document.characterSet, "windows-1252");
});
</script>
